Flipping around c'
Flipping around c' instead of around the last note. This has the advantage of creating far less errors due to notes being too high or too low. It has the disadvantage of the "hands" (e.g. bass and treble tracks) being swapped. That's why the old part of the program remains an option.